if click left button at blank space, it should select nothing
Add clearSelection() in the ListView::mousePressEvent() can slove this problem(#92).
and the reason why I add clearFocus() in ListView2::bindModel() is if I don't add this function then when I do this: 1.click an item; 2. click blank space; 3. click another item. The first item will be included in a dotted box while the menu pop up near the second item.