fix peony rename logic error
The main reason of these problem(#19,#79,#84) happens is because  when mouseDoubleClickEvent happens, it always trigger mousePressEvent first. It is hard to specify a single click and a double click.
So I add m_clickTimer to specify single click and double click. Only when single click happened then trigger rename slot.